Refactor graph generator
Changed how the edge ids are generated to avoid having different edges
treated as the same edge (ex 11,0 and 1,10 both give 110). Adding `-`
between the nodes id seems to fix the problem.

Also used a shuffled nodes array to avoid having to select a random target which
could produce a lot of misses when we get closer to a complete graph.