Refactor the code to use a graphology graph and add web workers support
Switched from sigma graph to graphology to enable the layout algorithm
to be run in the web workers without access to the DOM (required by
sigma). Until the edge events are added to sigam v2 we still need to
maintain a sigma graph instance to be able to render and interact with
the graph.

Other changes:
- Use webpack to handle dependencies and to enable the use of ES modules
in web workers.
- Only keep the build js instead of the full sigma source